Refactored: HTTP instrumentation into module.
